+++ src/apswversion.h
@@ -1,1 +1,1 @@
-#define APSW_VERSION "3.53.1.0"
+#define APSW_VERSION "3.53.2.0"